Do not modify event in-place on keypress callback.
When the `keyval` component was being modified in-place in the keypress callback
in order to allow keyboard shortcuts to work when caps lock is enabled, there was a
consequent effect that if the key pressed did not match any sakura shortcut, then
it would be upper cased definitely. This is fine unless the user has both caps lock
enabled *and* the Shift key pressed.

This fixes the issue by not modifying the GdkEventKey struct in-place in the callback.
Fixes Launchpad bug #1009261.